commit 3

+++ src/main/java/ru/kata/spring/boot_security/demo/repositories/UserRepository.java
@@ -15,7 +15,7 @@ public interface UserRepository {
 
     void updateUser(User user);
 
-    void removeUser(int id);
+    void removeUser(long id);
 
     User getUserById(long id);
 }
